+++ src/model/source-vim.ts
@@ -23,7 +23,6 @@ export default class VimSource extends Source {
   }
 
   public async shouldComplete(opt: CompleteOption): Promise<boolean> {
-    if (!this.checkFileType(opt.filetype)) return false
     if (this.optionalFns.indexOf('should_complete') === -1) return true
     let res = await this.callOptinalFunc('should_complete', [opt])
     return !!res