+++ src/source/include.ts
@@ -6,7 +6,6 @@ import Source from '../model/source'
 import {CompleteOption, CompleteResult, SourceConfig} from '../types'
 import {findSourceDir} from '../util/fs'
 import {toNumber} from '../util/types'
-import * as resolvers from './include_resolve'
 const exec = require('child_process').exec
 const logger = require('../util/logger')('source-include')
 const baseDir = path.join(__dirname, 'include_resolve')
@@ -33,16 +32,9 @@ export default class Include extends Source {
     }
   }
 
-  public async shouldComplete(opt: CompleteOption): Promise<boolean> {
-    let {filetype} = opt
-    if (!resolvers.hasOwnProperty(filetype)) return false
-    let {shouldResolve} = resolvers[filetype]
-    return await shouldResolve(opt)
-  }
-
   public async doComplete(opt: CompleteOption): Promise<CompleteResult> {
     let {command, nvim} = this
-    let {bufnr, col, input} = opt
+    let {bufnr, input} = opt
     if (input.length == 0) return null
     let {trimSameExts} = this.config
     let fullpath = await nvim.call('coc#util#get_fullpath', [toNumber(bufnr)])
@@ -71,7 +63,6 @@ export default class Include extends Source {
       }
     }
     return {
-      startcol: col - 1,
       items
     }
   }