+++ src/source/include_resolve/index.ts
@@ -1,9 +1,0 @@
-import * as javascript from './javascript'
-import * as typescript from './typescript'
-import * as wxml from './wxml'
-
-export {
-  javascript,
-  typescript,
-  wxml
-}