+++ src/source/word.ts
@@ -18,7 +18,6 @@ export default class Word extends Source {
   }
 
   public async shouldComplete(opt: CompleteOption): Promise<boolean> {
-    if (!this.checkFileType(opt.filetype)) return false
     let {input} = opt
     if (input.length === 0) return false
     return true